Growing {V} for adj list is not worth it.
This commit fixes the adjacency list as an array of lists.

Each index is a vertex and the off shooting list is the adjacent
vertices.  The justification for fixing this is because growing
{V} should not be common (right? -- at least compared with {E}.

Also fixed up the linked list DS and tests a bit.  It should be
less broken (syntastic does not complain much any more).